+++ ruma-events/tests/pdu.rs
@@ -16,8 +16,10 @@ use serde_json::{from_value as from_json_value, json, to_value as to_json_value}
 fn serialize_stub_as_v1() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
     signatures.insert("example.com".try_into().unwrap(), inner_signature);
 
     let mut unsigned = BTreeMap::new();
@@ -77,8 +79,10 @@ fn serialize_stub_as_v1() {
 fn serialize_stub_as_v3() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
     signatures.insert("example.com".try_into().unwrap(), inner_signature);
 
     let mut unsigned = BTreeMap::new();
@@ -158,7 +162,7 @@ fn deserialize_stub_as_v1() {
         "sender": "@someone:matrix.org",
         "signatures": {
             "example.com": {
-                "ed25519:key_version:": "86BytesOfSignatureOfTheRedactedEvent"
+                "ed25519:key_version": "86BytesOfSignatureOfTheRedactedEvent"
             }
         },
         "state_key": "my_key",
@@ -205,7 +209,7 @@ fn deserialize_stub_as_v3() {
         "sender": "@someone:matrix.org",
         "signatures": {
             "example.com": {
-                "ed25519:key_version:": "86BytesOfSignatureOfTheRedactedEvent"
+                "ed25519:key_version": "86BytesOfSignatureOfTheRedactedEvent"
             }
         },
         "state_key": "my_key",
@@ -228,8 +232,10 @@ fn deserialize_stub_as_v3() {
 fn serialize_pdu_as_v1() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
     signatures.insert("example.com".try_into().unwrap(), inner_signature);
 
     let mut unsigned = BTreeMap::new();
@@ -293,8 +299,10 @@ fn serialize_pdu_as_v1() {
 fn serialize_pdu_as_v3() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
     signatures.insert("example.com".try_into().unwrap(), inner_signature);
 
     let mut unsigned = BTreeMap::new();
@@ -378,7 +386,7 @@ fn test_deserialize_pdu_as_v1() {
         "sender": "@someone:matrix.org",
         "signatures": {
             "example.com": {
-                "ed25519:key_version:": "86BytesOfSignatureOfTheRedactedEvent"
+                "ed25519:key_version": "86BytesOfSignatureOfTheRedactedEvent"
             }
         },
         "state_key": "my_key",
@@ -426,7 +434,7 @@ fn deserialize_pdu_as_v3() {
         "sender": "@someone:matrix.org",
         "signatures": {
             "example.com": {
-                "ed25519:key_version:": "86BytesOfSignatureOfTheRedactedEvent"
+                "ed25519:key_version": "86BytesOfSignatureOfTheRedactedEvent"
             }
         },
         "state_key": "my_key",
@@ -449,8 +457,10 @@ fn deserialize_pdu_as_v3() {
 fn convert_v1_stub_to_pdu() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
     signatures.insert("example.com".try_into().unwrap(), inner_signature);
 
     let mut unsigned = BTreeMap::new();
@@ -523,8 +533,10 @@ fn convert_v1_stub_to_pdu() {
 fn convert_v3_stub_to_pdu() {
     let mut signatures = BTreeMap::new();
     let mut inner_signature = BTreeMap::new();
-    inner_signature
-        .insert("ed25519:key_version".into(), "86BytesOfSignatureOfTheRedactedEvent".into());
+    inner_signature.insert(
+        "ed25519:key_version".try_into().unwrap(),
+        "86BytesOfSignatureOfTheRedactedEvent".into(),
+    );
 
     signatures.insert("example.com".try_into().unwrap(), inner_signature);